backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vidpipe"
version = "0.1.0"
description = "AI-powered video pipeline for multi-scene video generation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Video Pipeline Team"}
]
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.22.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "typer>=0.12.0",
    "rich>=13.0",
    "Pillow>=10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["vidpipe*"]
